Move file helper classes into the files module
