+++ wagtail_transfer/files.py
@@ -1,6 +1,12 @@
 import hashlib
 from contextlib import contextmanager
 
+from django.core.files.base import ContentFile
+
+import requests
+
+from .models import ImportedFile
+
 
 @contextmanager
 def open_file(field, file):
@@ -79,3 +85,36 @@ def get_file_hash(field, instance):
     # Fall back to calculating it on the fly
     with open_file(field, field.value_from_object(instance)) as f:
         return hashlib.sha1(f.read()).hexdigest()
+
+
+class FileTransferError(Exception):
+    pass
+
+
+class File:
+    """
+    Represents a file that needs to be imported
+
+    Note that local_filename is only a guideline, it may be changed to avoid conflict with an existing file
+    """
+    def __init__(self, local_filename, size, hash, source_url):
+        self.local_filename = local_filename
+        self.size = size
+        self.hash = hash
+        self.source_url = source_url
+
+    def transfer(self):
+        response = requests.get(self.source_url)
+
+        if response.status_code != 200:
+            raise FileTransferError  # TODO
+
+        return ImportedFile.objects.create(
+            file=ContentFile(response.content, name=self.local_filename),
+            source_url=self.source_url,
+            hash=self.hash,
+            size=self.size,
+        )
+
+    def __hash__(self):
+        return hash((self.local_filename, self.size, self.hash, self.source_url))
